version 453

+++ dayrui/Fcms/Model/Module.php
@@ -312,16 +312,18 @@ class Module extends \Phpcmf\Model
             return dr_return_data(0, dr_lang('此模块属于应用类型，请到[本地应用]中去卸载'));
         }
 
-        $row = $this->db->table('member_setting')->where('name', 'auth_module')->get()->getRowArray();
-        $auth_module = dr_string2array($row['value']);
         if (isset($site[SITE_ID]) && $site[SITE_ID]) {
             // 删除当前站点中的全部模块表
-            if (isset($auth_module[SITE_ID][$dir])) {
-                unset($auth_module[SITE_ID][$dir]);
-                $this->db->table('member_setting')->replace([
-                    'name' => 'auth_module',
-                    'value' => dr_array2string($auth_module)
-                ]);
+            if (IS_USE_MEMBER) {
+                $row = $this->db->table('member_setting')->where('name', 'auth_module')->get()->getRowArray();
+                $auth_module = dr_string2array($row['value']);
+                if (isset($auth_module[SITE_ID][$dir])) {
+                    unset($auth_module[SITE_ID][$dir]);
+                    $this->db->table('member_setting')->replace([
+                        'name' => 'auth_module',
+                        'value' => dr_array2string($auth_module)
+                    ]);
+                }
             }
             // 系统模块
             // 删除系统表
@@ -401,7 +403,7 @@ class Module extends \Phpcmf\Model
             $this->db->table('field')->where('relatedid', $module['id'])->where('relatedname', 'module')->delete();
             // 删除菜单
             $this->db->table('admin_menu')->like('mark', 'module-'.$dir)->delete();
-            $this->db->table('member_menu')->like('mark', 'module-'.$dir)->delete();
+            IS_USE_MEMBER && $this->db->table('member_menu')->like('mark', 'module-'.$dir)->delete();
             // 删除自定义菜单
             \Phpcmf\Service::M('Menu')->delete_app($dir);
             // 删除自定义表单
